// YellowPage.Client/src/Components/AddPersonForm/types.ts
export type AddNewPersonFormType = {
  firstName: { value: string };
  lastName: { value: string };
  email: { value: string };
  phoneNumber: { value: string };
  webSite: { value: string };
  country: { value: string };
  city: { value: string };
  address: { value: string };
  zipCode: { value: string };
};